+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 * Add an overview of all active plugins in the settings about section [CTC-1667]
 * Add an archetype for xill plugins [CTC-1284]
 * Make non-robot files visible in the Xill IDE project pane [CTC-1652]
+* Make non-robot files editable in Xill IDE [CTC-1617]
 
 ### Change